w3c-reports/12092014/mac-wk2/runned-tests/workers/semantics/multiple-workers/005-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: SharedWorker

Harness Error. harness_status.status = 1 , harness_status.message = ReferenceError: Can't find variable: SharedWorker
